+++ build.js
@@ -34,13 +34,13 @@ const esbuildConfig = {
   entryPoints: [
     "./src/content.js",
     "./src/popup.js",
+    "./src/background.js",
     "./src/inject_script.js",
-    "./src/browser_panel/panel.js",
-    "./src/scripts/backend.js",
-    "./src/scripts/background.js",
-    "./src/scripts/content_proxy.js",
-    "./src/scripts/devtools-background.js",
-    "./src/scripts/proxy.js",
+    "./src/browser_panel/panel/panel.js",
+    "./src/browser_panel/panel/register.js",
+    "./src/browser_panel/page/backend.js",
+    "./src/browser_panel/page/content_proxy.js",
+    "./src/browser_panel/proxy.js",
   ],
   bundle: true,
   minify: production,
