+++ public/panel.html
@@ -105,6 +105,6 @@
       </div>
     </div>
 
-    <script src="dist/browser_panel/panel.js" type="module"></script>
+    <script src="dist/browser_panel/panel/panel.js" type="module"></script>
   </body>
 </html>
