+++ public/register.html
@@ -1,0 +1,9 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+  </head>
+  <body>
+    <script src="./dist/browser_panel/panel/register.js"></script>
+  </body>
+</html>
