+++ src/browser_panel/panel/register.js
