+++ src/browser_panel/proxy.js
@@ -2,7 +2,7 @@
 // activated. Because it is not injected using eval, it has full privilege
 // to the chrome runtime API. It serves as a proxy between the injected
 // backend and the DevTool panel.
-import { HOTWIRE_DEV_TOOLS_BACKEND_SOURCE, HOTWIRE_DEV_TOOLS_PROXY_SOURCE, PROXY } from "../browser_panel/ports"
+import { HOTWIRE_DEV_TOOLS_BACKEND_SOURCE, HOTWIRE_DEV_TOOLS_PROXY_SOURCE, PROXY } from "./ports"
 
 function proxy() {
   const proxyPort = chrome.runtime.connect({
